LPD-42546 Add missing schema prop
