+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/resources/META-INF/resources/js/CustomInternalViewPropsTransformer.ts
@@ -13,6 +13,14 @@ export default function propsTransformer({...otherProps}: any) {
 		default: true,
 		label: 'My Carousel View',
 		name: 'carouselViewRenderer',
+		schema: {
+			description: 'description',
+			image: '',
+			link: '',
+			sticker: '',
+			symbol: '',
+			title: 'title',
+		},
 		symbol: 'rotate',
 		type: 'internal',
 	};
