+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/FrontendDataSet.js
@@ -123,6 +123,7 @@ const FrontendDataSet = ({
 				default: customRenderer.default,
 				label: customRenderer.label,
 				name: customRenderer.name,
+				schema: customRenderer.schema,
 				thumbnail: customRenderer.symbol,
 			})) || [];
